fix: need to authorizer if user is authorized for current collection + set tenant & venue for DB

+++ cumulus_lambda_functions/uds_api/granules_api.py
@@ -46,7 +46,21 @@ async def dapa_archive_add_config(request: Request, collection_id: str, new_body
 @router.post("/{collection_id}/archive/")
 async def dapa_archive_update_config(request: Request, collection_id: str, new_body: DaacAddModel):
     LOGGER.debug(f'started dapa_archive_add_config. {new_body.model_dump()}')
+    authorizer: UDSAuthorizorAbstract = UDSAuthorizerFactory() \
+        .get_instance(UDSAuthorizerFactory.cognito,
+                      es_url=os.getenv('ES_URL'),
+                      es_port=int(os.getenv('ES_PORT', '443'))
+                      )
     auth_info = FastApiUtils.get_authorization_info(request)
+    collection_identifier = UdsCollections.decode_identifier(collection_id)
+    if not authorizer.is_authorized_for_collection(DBConstants.read, collection_id,
+                                                   auth_info['ldap_groups'],
+                                                   collection_identifier.tenant,
+                                                   collection_identifier.venue):
+        LOGGER.debug(f'user: {auth_info["username"]} is not authorized for {collection_id}')
+        raise HTTPException(status_code=403, detail=json.dumps({
+            'message': 'not authorized to execute this action'
+        }))
     daac_crud = DaacArchiveCrud(auth_info, collection_id, new_body.model_dump())
     add_result = daac_crud.update_config()
     if add_result['statusCode'] == 200:
@@ -57,7 +71,21 @@ async def dapa_archive_update_config(request: Request, collection_id: str, new_b
 @router.delete("/{collection_id}/archive/")
 async def dapa_archive_delete_config(request: Request, collection_id: str, new_body: DaacDeleteModel):
     LOGGER.debug(f'started dapa_archive_add_config. {new_body.model_dump()}')
+    authorizer: UDSAuthorizorAbstract = UDSAuthorizerFactory() \
+        .get_instance(UDSAuthorizerFactory.cognito,
+                      es_url=os.getenv('ES_URL'),
+                      es_port=int(os.getenv('ES_PORT', '443'))
+                      )
     auth_info = FastApiUtils.get_authorization_info(request)
+    collection_identifier = UdsCollections.decode_identifier(collection_id)
+    if not authorizer.is_authorized_for_collection(DBConstants.read, collection_id,
+                                                   auth_info['ldap_groups'],
+                                                   collection_identifier.tenant,
+                                                   collection_identifier.venue):
+        LOGGER.debug(f'user: {auth_info["username"]} is not authorized for {collection_id}')
+        raise HTTPException(status_code=403, detail=json.dumps({
+            'message': 'not authorized to execute this action'
+        }))
     daac_crud = DaacArchiveCrud(auth_info, collection_id, new_body.model_dump())
     add_result = daac_crud.delete_config()
     if add_result['statusCode'] == 200:
@@ -68,7 +96,21 @@ async def dapa_archive_delete_config(request: Request, collection_id: str, new_b
 @router.get("/{collection_id}/archive/")
 async def dapa_archive_get_config(request: Request, collection_id: str):
     # TODO return UDS SNS to accept DAAC messages here
+    authorizer: UDSAuthorizorAbstract = UDSAuthorizerFactory() \
+        .get_instance(UDSAuthorizerFactory.cognito,
+                      es_url=os.getenv('ES_URL'),
+                      es_port=int(os.getenv('ES_PORT', '443'))
+                      )
     auth_info = FastApiUtils.get_authorization_info(request)
+    collection_identifier = UdsCollections.decode_identifier(collection_id)
+    if not authorizer.is_authorized_for_collection(DBConstants.read, collection_id,
+                                                   auth_info['ldap_groups'],
+                                                   collection_identifier.tenant,
+                                                   collection_identifier.venue):
+        LOGGER.debug(f'user: {auth_info["username"]} is not authorized for {collection_id}')
+        raise HTTPException(status_code=403, detail=json.dumps({
+            'message': 'not authorized to execute this action'
+        }))
     daac_crud = DaacArchiveCrud(auth_info, collection_id, {})
     add_result = daac_crud.get_config()
     if add_result['statusCode'] == 200:
